requirements.txt
elasticsearch==8.3.3
fastapi==0.88.0
nltk==3.7
numpy==1.23.5
pandas==1.4.3
PyYAML==6.0
scikit-learn==1.1.3
scipy==1.9.3
Send2Trash==1.8.0
sentence-transformers==2.2.2
transformers==4.25.1

